[add] header and description styles
